pyproject.toml
[tool.poetry]
authors = ["Arash Hatami <hatamiarash7@gmail.com>"]
description = "Update blog using Telegram messages"
name = "tg-blog-updater"
packages = [
  {include = "tg_blog_updater", from = "."},
]
readme = "README.md"
version = "0.2.0"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.1.1"
pygithub = "^2.2.0"
pillow = "^11.2.1"
python-slugify = {extras = ["unidecode"], version = "^8.0.4"}

[tool.poetry.group.test.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.26.0"

[tool.poetry.group.dev.dependencies]
pylint = "^3.0.0"
pytest = "^8.0.0"
pytest-asyncio = "^0.26.0"
isort = "^5.13.2"
black = "^24.4.2"

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/hatamiarash7/tg-blog-updater/issues"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
target-version = "py312"
